refactor(backend): move listennotes utils to module
