+++ apps/backend/src/utils.rs
@@ -191,41 +191,6 @@ pub fn get_now_timestamp() -> u128 {
         .as_millis()
 }
 
-pub mod listennotes {
-    use std::collections::HashMap;
-
-    use crate::graphql::PROJECT_NAME;
-
-    use super::*;
-
-    pub async fn get_client_config(url: &str, api_token: &str) -> (Client, HashMap<i32, String>) {
-        let client: Client = Config::new()
-            .add_header("X-ListenAPI-Key", api_token)
-            .unwrap()
-            .add_header(USER_AGENT, format!("{}/{}", AUTHOR, PROJECT_NAME))
-            .unwrap()
-            .set_base_url(Url::parse(url).unwrap())
-            .try_into()
-            .unwrap();
-        #[derive(Debug, Serialize, Deserialize, Default)]
-        struct Genre {
-            id: i32,
-            name: String,
-        }
-        #[derive(Debug, Serialize, Deserialize, Default)]
-        struct GenreResponse {
-            genres: Vec<Genre>,
-        }
-        let mut rsp = client.get("genres").await.unwrap();
-        let data: GenreResponse = rsp.body_json().await.unwrap_or_default();
-        let mut genres = HashMap::new();
-        for genre in data.genres {
-            genres.insert(genre.id, genre.name);
-        }
-        (client, genres)
-    }
-}
-
 pub mod igdb {
     use std::{env, fs};
